Clean up some stuff in jitsi video (#130)
