+++ app/client/jsx/JitsiVideo.jsx
@@ -4,6 +4,7 @@ import reducers from './reducers.jsx'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faSpinner } from '@fortawesome/free-solid-svg-icons'
 import Config from './Config.jsx'
+import { transparentJitsi } from './utils.js'
 
 class JitsiVideo extends Component {
     constructor(props) {
@@ -61,43 +62,12 @@ class JitsiVideo extends Component {
         iframe.setAttribute('allowtransparency', 'true')
         try {
             const iframeDocument = iframe.contentWindow.document
-            const css = `
-                body {
-                    background-color: transparent !important;
-                }
-                body #largeVideoContainer.videoContainer {
-                    background-color: transparent !important;
-                }
-                .tOoji {
-                    background-color: transparent !important;
-                }
-            `
             const style = iframeDocument.createElement('style')
             iframeDocument.head.appendChild(style);
             style.type = 'text/css';
-            style.appendChild(iframeDocument.createTextNode(css));
+            style.appendChild(iframeDocument.createTextNode(transparentJitsi.css));
         } catch (err) {
-            console.log("%c******************************************************", 'background: #222; color: #bada55')
-            console.log("Can't make Jitsi transparent! Must be running locally.")
-            console.log("To test the background, inspect the document inside the iframe and paste this code into the console:")
-            console.log(`
-const css = \`
-body {
-    background-color: transparent !important;
-}
-body #largeVideoContainer.videoContainer {
-    background-color: transparent !important;
-}
-.tOoji {
-    background-color: transparent !important;
-}
-\`
-const style = document.createElement('style')
-document.head.appendChild(style);
-style.type = 'text/css';
-style.appendChild(document.createTextNode(css));
-            `)
-            console.log("%c******************************************************", 'background: #222; color: #bada55')
+            transparentJitsi.printInstructions()
         }
     }
 
@@ -115,7 +85,6 @@ style.appendChild(document.createTextNode(css));
             }
 
             const domain = 'party.gbre.org/jitsi/'
-            // const domain = 'jitsi.gbre.org'
             // const domain = 'meet.jit.si';
             const options = {
                 roomName: this.props.jitsiData.roomName,
@@ -141,7 +110,7 @@ style.appendChild(document.createTextNode(css));
                 const avatarUrl = Config.avatars[type][color]
                 const commands = {
                     displayName: this.props.jitsiData.displayName,
-                    avatarUrl,
+                    avatarUrl
                 }
                 // Persist audio/video muted settings unless acquired by muteRoom room setting
                 if (this.isAudioMuted || this.props.jitsiData.muteRoom) {
